Refactored entity match schemas
