+++ server/graphql/schema/common/index.ts
@@ -2,4 +2,4 @@ import enums from "./enums";
 import inputs from "./inputs";
 import types from "./types";
 
-export default [enums, types, inputs];
+export default [enums, inputs, types];
