+++ server/graphql/schema/common/inputs.ts
@@ -5,35 +5,4 @@ export default gql`
     field: OrderField
     direction: OrderDirection
   }
-
-  # --------- Match -------------
-
-  # ------- matchProjectsToMember ------
-  input matchProjectsToMemberInput {
-    memberID: ID!
-    serverID: [ID]
-  }
-
-  # ------- matchMembersToProject ------
-  input matchMembersToProjectInput {
-    projectID: ID!
-    serverID: [ID]
-  }
-
-  # ------- matchMembersToProject ------
-
-  input matchMembersToProjectRoleInput {
-    projectRoleID: ID!
-    serverID: [ID]
-  }
-
-  input matchMembersToSkillsInput {
-    skillIDs: [ID]!
-    serverID: [ID]
-  }
-
-  input matchMembersToUserInput {
-    memberID: ID!
-    serverID: [ID]
-  }
 `;
