+++ server/graphql/schema/common/types.ts
@@ -28,39 +28,4 @@ export default gql`
     start: String
     end: String
   }
-
-  type matchProjectsCursorOutput {
-    edges: [ProjectMatchEdge]
-    pageInfo: PageInfo
-  }
-
-  type ProjectMatchEdge {
-    cursor: String
-    matchProject: MatchProjectInfo
-    matchProjectRoles: [MatchProjectRole]
-  }
-
-  type MatchProjectInfo {
-    project: Project
-    matchPercentage: Int
-    relatedSkills: [Skill]
-  }
-
-  type MatchProjectRole {
-    role: ProjectRole
-    matchPercentage: Int
-    relatedSkills: [Skill]
-  }
-
-  type matchMembersCursorOutput {
-    edges: [MemberMatchEdge]
-    pageInfo: PageInfo
-  }
-
-  type MemberMatchEdge {
-    cursor: String
-    member: Member
-    matchPercentage: Int
-    relatedSkills: [Skill]
-  }
 `;
