+++ server/graphql/schema/index.ts
@@ -1,4 +1,3 @@
-import query from "./query";
 import memberType from "./member";
 import serverType from "./server";
 import skillType from "./skill";
@@ -7,9 +6,9 @@ import projectType from "./project";
 import commonType from "./common";
 import skillCategory from "./skillCategory";
 import rooms from "./rooms";
+import matches from "./matches";
 
 const typeDefs = [
-  query,
   ...memberType,
   ...serverType,
   ...skillType,
@@ -18,6 +17,7 @@ const typeDefs = [
   ...commonType,
   ...skillCategory,
   ...rooms,
+  ...matches,
 ];
 
 export default typeDefs;
