+++ server/graphql/schema/matches/index.ts
@@ -1,0 +1,5 @@
+import inputs from "./inputs";
+import query from "./query";
+import types from "./types";
+
+export default [inputs, types, query];
